hotfix : 회원 탈퇴 API 수정

+++ src/main/java/com/sunny/backend/apple/service/AppleService.java
@@ -11,6 +11,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
 import org.bouncycastle.openssl.jcajce.JcaPEMKeyConverter;
@@ -30,13 +32,15 @@ import com.sunny.backend.common.exception.CustomException;
 import com.sunny.backend.common.response.CommonResponse;
 import com.sunny.backend.common.response.ResponseService;
 import com.sunny.backend.competition.repository.CompetitionRepository;
-import com.sunny.backend.friends.domain.Friend;
 import com.sunny.backend.friends.domain.FriendCompetition;
 import com.sunny.backend.friends.repository.FriendCompetitionRepository;
 import com.sunny.backend.friends.repository.FriendRepository;
 import com.sunny.backend.notification.repository.CommentNotificationRepository;
+import com.sunny.backend.notification.repository.CompetitionNotificationRepository;
 import com.sunny.backend.notification.repository.FriendsNotificationRepository;
 import com.sunny.backend.notification.repository.NotificationRepository;
+import com.sunny.backend.report.repository.CommentReportRepository;
+import com.sunny.backend.report.repository.CommunityReportRepository;
 import com.sunny.backend.user.domain.Users;
 import com.sunny.backend.user.repository.UserRepository;
 import com.sunny.backend.util.RedisUtil;
@@ -63,6 +67,9 @@ public class AppleService {
 	private final ResponseService responseService;
 	private final RedisUtil redisUtil;
 	private final FriendCompetitionRepository friendCompetitionRepository;
+	private final CompetitionNotificationRepository competitionNotificationRepository;
+	private final CommentReportRepository commentReportRepository;
+	private final CommunityReportRepository communityReportRepository;
 
 	private final TokenProvider tokenProvider;
 
@@ -137,18 +144,22 @@ public class AppleService {
 				log.info("Apple token 삭제 성공");
 				notificationRepository.deleteByUsers(users);
 
-				for (Friend friend : friendRepository.findByUsers(users)) {
-					List<FriendCompetition> friendCompetitions = friendCompetitionRepository.getByFriendAndCompetition(
-						friend.getId(), null);
-					friendCompetitionRepository.deleteAllByFriend(friend);
-					List<Long> competitionIds = friendCompetitions.stream()
-						.map(friendCompetition -> friendCompetition.getCompetition().getId())
-						.toList();
+				List<FriendCompetition> friendCompetitions = friendCompetitionRepository.getByUserOrUserFriend(userId);
+				for (FriendCompetition friendCompetition : friendCompetitions) {
+					competitionNotificationRepository.deleteAllByFriendCompetition(friendCompetition);
+					friendCompetitionRepository.deleteById(friendCompetition.getId());
+				}
+				Set<Long> competitionIds = friendCompetitions.stream()
+					.map(friendCompetition -> friendCompetition.getCompetition().getId())
+					.collect(Collectors.toSet());
+				if (!competitionIds.isEmpty()) {
 					competitionRepository.deleteAllById(competitionIds);
 				}
+				
 				friendsNotificationRepository.deleteByUsersOrFriend(users, users);
 				friendRepository.deleteByUsersOrUserFriend(users, users);
-				commentNotificationRepository.deleteByUsers(users);
+				communityReportRepository.deleteAllByCommunityInOrUsers(users.getCommunityList(), users);
+				commentReportRepository.deleteAllByCommentInOrUsers(users.getCommentList(), users);
 				commentRepository.nullifyUsersId(users.getId());
 				userRepository.deleteById(users.getId());
 				log.info("Apple token 삭제 성공 code={}", HttpStatus.OK.value());
